+++ docs/fr_FR/changelog.md
@@ -1,1 +1,7 @@
-Version 0.1.0
+# Changelog JeedomConnect
+## Version 0.5.0
+- Ajout d'un widget Mode
+- Ajout d'un widget Action Générique (other)
+
+## Version 0.1.0
+Version initiale